+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 ### Fixed
 - \#3160 - Show server-side validation errors for invalid constraints
 - \#3164 - Show server-side validation errors for invalid object
+- \#3214 -  App creation modal ports field default value does not reflect
+  service default value
+- \#3218 - Regression: group name no longer pre-filled in app modal
 
 ## 0.15.3 - 2016-02-03
 ### Fixed